# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API_GET_POST_UPDATE _DELETE_DATA",
        "description": "API PROJECT MODEL CHANGE RECORD",
        "version": "1.0.0"
    },
    "paths": {
...
        "/oprform/{id}": {
            "put": {
                "tags": [
                    "OPR CP Form"
                ],
                "summary": "Update OPR Process CP form data",
                "operationId": "73b11cf5136e19faf05dd585b1ee2672",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Record ID to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "prs": {
                                        "type": "string",
                                        "example": "CP"
                                    },
                                    "ecn": {
                                        "type": "string",
                                        "example": "Yes"
                                    },
                                    "ecn_num": {
                                        "type": "string",
                                        "example": "E1234"
                                    },
                                    "ecn_rev": {
                                        "type": "string",
                                        "example": "01"
                                    },
                                    "date_rec": {
                                        "type": "string",
                                        "example": "2025-10-08"
                                    },
                                    "line": {
                                        "type": "string",
                                        "example": "Line A"
                                    },
                                    "issue_no": {
                                        "type": "string",
                                        "example": "IS-2025-001"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "lots": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "prog_name": {
                                        "type": "string",
                                        "example": "Program X"
                                    },
                                    "prog_rev": {
                                        "type": "string",
                                        "example": "Rev A"
                                    },
                                    "empno": {
                                        "type": "string",
                                        "example": "2970011"
                                    },
                                    "cur_modelname": {
                                        "type": "string",
                                        "example": "Model-CUR"
                                    },
                                    "chn_modelname": {
                                        "type": "string",
                                        "example": "Model-CHN"
                                    },
                                    "won_cur": {
                                        "type": "string",
                                        "example": "WO-CUR-001"
                                    },
                                    "won_chn": {
                                        "type": "string",
                                        "example": "WO-CHN-002"
                                    },
                                    "lots_chn": {
                                        "type": "string",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/oprform/rf/{id}": {
            "put": {
                "tags": [
                    "OPR RF Form"
                ],
                "summary": "Update OPR Process RF form data",
                "operationId": "ceddbff6f8c84cbdffc0332ef1e0ba23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Record ID to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "prs": {
                                        "type": "string"
                                    },
                                    "procs_rf": {
                                        "type": "string"
                                    },
                                    "ecn": {
                                        "type": "string"
                                    },
                                    "ecn_num": {
                                        "type": "string"
                                    },
                                    "ecn_rev": {
                                        "type": "string"
                                    },
                                    "date_rec": {
                                        "type": "string"
                                    },
                                    "line": {
                                        "type": "string"
                                    },
                                    "issue_no": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "lots": {
                                        "type": "number"
                                    },
                                    "prog_name": {
                                        "type": "string"
                                    },
                                    "prog_rev": {
                                        "type": "string"
                                    },
                                    "empno": {
                                        "type": "string"
                                    },
                                    "cur_modelname": {
                                        "type": "string"
                                    },
                                    "chn_modelname": {
                                        "type": "string"
                                    },
                                    "won_cur": {
                                        "type": "string"
                                    },
                                    "won_chn": {
                                        "type": "string"
                                    },
                                    "lots_chn": {
                                        "type": "number"
                                    },
                                    "procs_rf_chn": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/data/delete/{id}": {
            "delete": {
                "tags": [
                    "OPR Form"
                ],
                "summary": "Delete OPR form by ID",
                "operationId": "5e51ba02f7e1a935b1da76231ad2e86b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the record to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/submit/{id}": {
            "put": {
                "tags": [
                    "OPR Form"
                ],
                "summary": "Submit OPR form by ID",
                "operationId": "b2026c950001d2f12abefba30994bfbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the record to submit",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        }
    },
...
}